refactor(api): move token controller
to token directory
